+++ App/Helpers/Helper.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Text;
+
+namespace App.Helpers
+{
+    public static class Helper
+    {
+        public static string GenerateRandomString(int length)
+        {
+            Random random = new Random();
+            string characters = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
+            StringBuilder result = new StringBuilder(length);
+            for (int i = 0; i < length; i++)
+            {
+                result.Append(characters[random.Next(characters.Length)]);
+            }
+
+            return result.ToString();
+        }
+    }
+}
